flutterapi/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
Django==3.2
django-filter==2.4.0
djangorestframework==3.12.4
Flask==2.0.2
Flask-WTF==0.15.1
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
joblib==1.1.0
Markdown==3.3.4
MarkupSafe==2.0.1
numpy==1.21.3
opencv-python==4.5.4.58
pytz==2021.1
requests==2.26.0
scikit-learn==1.0.1
scipy==1.7.1
sklearn==0.0
sqlparse==0.4.2
threadpoolctl==3.0.0
torch==1.10.0
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
WTForms==2.3.3
